PP-14150-Replace unused queue directory with the updated queue-dropwizard-4-aws-sdk-v2
Replace and remove redundant queue-dropwizard-4-aws-sdk-v2 modules
