--- queue-dropwizard-4-aws-sdk-v2/pom.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <artifactId>pay-java-commons</artifactId>
-        <groupId>uk.gov.service.payments</groupId>
-        <version>1.0.0</version>
-    </parent>
-
-    <artifactId>queue-dropwizard-4-aws-sdk-v2</artifactId>
-    <name>Classes for receiving and sending queue messages DW 4</name>
-    <packaging>jar</packaging>
-
-    <dependencies>
-        <!-- Main dependencies that are imported from one of the BOMs specified
-             in <dependencyManagement> in the parent POM, so no explicit versions needed -->
-        <dependency>
-            <groupId>io.dropwizard</groupId>
-            <artifactId>dropwizard-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.dropwizard</groupId>
-            <artifactId>dropwizard-json-logging</artifactId>
-        </dependency>
-
-        <!-- Main dependencies that need explicit versions -->
-        <dependency>
-            <groupId>software.amazon.awssdk</groupId>
-            <artifactId>sqs</artifactId>
-            <version>2.32.24</version>
-        </dependency>
-
-        <!-- Test dependencies that are imported from one of the BOMs specified
-             in <dependencyManagement> in the parent POM, so no explicit versions needed -->
-
-        <!-- Test dependencies that need explicit versions -->
-        <dependency>
-            <groupId>org.hamcrest</groupId>
-            <artifactId>hamcrest</artifactId>
-            <version>${hamcrest.version}</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.hamcrest</groupId>
-            <artifactId>hamcrest-library</artifactId>
-            <version>${hamcrest.version}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-</project>
