--- queue-dropwizard-4-aws-sdk-v2/src/main/java/uk/gov/service/payments/commons/queue/exception/QueueException.java
@@ -1,12 +1,0 @@
-package uk.gov.service.payments.commons.queue.exception;
-
-public class QueueException extends Exception {
-
-    public QueueException(){
-        
-    }
-    
-    public QueueException(String message) {
-        super(message);
-    }
-}
