--- queue-dropwizard-4-aws-sdk-v2/src/main/java/uk/gov/service/payments/commons/queue/sqs/AbstractQueue.java
@@ -1,46 +1,0 @@
-package uk.gov.service.payments.commons.queue.sqs;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import uk.gov.service.payments.commons.queue.exception.QueueException;
-import uk.gov.service.payments.commons.queue.model.QueueMessage;
-
-import java.util.List;
-
-public abstract class AbstractQueue {
-
-    private static final String MESSAGE_ATTRIBUTES_TO_RECEIVE = "All";
-    protected ObjectMapper objectMapper;
-    private String queueUrl;
-    private int failedMessageRetryDelayInSeconds;
-    private SqsQueueService sqsQueueService;
-
-    public AbstractQueue(SqsQueueService sqsQueueService, ObjectMapper objectMapper,
-                         String queueUrl, int failedMessageRetryDelayInSeconds) {
-        this.sqsQueueService = sqsQueueService;
-        this.queueUrl = queueUrl;
-        this.failedMessageRetryDelayInSeconds = failedMessageRetryDelayInSeconds;
-        this.objectMapper = objectMapper;
-    }
-
-    public QueueMessage sendMessageToQueue(String message) throws QueueException {
-        return sqsQueueService.sendMessage(queueUrl, message);
-    }
-
-    public QueueMessage sendMessageToQueueWithDelay(String message, int delayInSeconds) throws QueueException {
-        return sqsQueueService.sendMessage(queueUrl, message, delayInSeconds);
-    }
-
-    public List<QueueMessage> retrieveMessages() throws QueueException {
-        return sqsQueueService
-                .receiveMessages(this.queueUrl, MESSAGE_ATTRIBUTES_TO_RECEIVE);
-    }
-
-    public void markMessageAsProcessed(QueueMessage queueMessage) throws QueueException {
-        sqsQueueService.deleteMessage(this.queueUrl, queueMessage.getReceiptHandle());
-    }
-
-    public void scheduleMessageForRetry(QueueMessage queueMessage) throws QueueException {
-        sqsQueueService.deferMessage(this.queueUrl, queueMessage.getReceiptHandle(), failedMessageRetryDelayInSeconds);
-    }
-}
-
