+++ queue-dropwizard-4/pom.xml
@@ -27,9 +27,9 @@
 
         <!-- Main dependencies that need explicit versions -->
         <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk-sqs</artifactId>
-            <version>1.12.782</version>
+            <groupId>software.amazon.awssdk</groupId>
+            <artifactId>sqs</artifactId>
+            <version>2.32.24</version>
         </dependency>
 
         <!-- Test dependencies that are imported from one of the BOMs specified
